jenkins/custom_tools/MetricsProcessor/src/main/java/com/infosys/json/FxCopReportJson.java
/***********************************************************************************************
*
* Copyright 2018 Infosys Ltd.
* Use of this source code is governed by MIT license that can be found in the LICENSE file or at
* https://opensource.org/licenses/MIT.
*
***********************************************************************************************/
package com.infosys.json;

import com.google.gson.annotations.Expose;
import com.google.gson.annotations.SerializedName;

public class FxCopReportJson {
	@SerializedName("high")
    @Expose
    private Integer high;
	@SerializedName("medium")
    @Expose
    private Integer medium;
	@SerializedName("low")
    @Expose
    private Integer low;
	public Integer getHigh() {
		return high;
	}
	public void setHigh(Integer high) {
		this.high = high;
	}
	public Integer getMedium() {
		return medium;
	}
	public void setMedium(Integer medium) {
		this.medium = medium;
	}
	public Integer getLow() {
		return low;
	}
	public void setLow(Integer low) {
		this.low = low;
	}
	
}
